% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawpathfinder.R
\name{rawpathfinder}
\alias{rawpathfinder}
\title{Query rawpathfinder for files/folders raw unix paths.}
\usage{
rawpathfinder(query, protocol = "http", ip = "192.168.1.209", port = 8958)
}
\arguments{
\item{query}{A character vector with names of files/folders.}

\item{protocol}{Used protocol: http or https.}

\item{ip}{The ip of the rawpathfinder flask service.}

\item{port}{The port of the rawpathfinder flask service.}
}
\value{
A data frame with files/folder names maped to unix paths.
}
\description{
Query rawpathfinder for files/folders raw unix paths.
}
\examples{
\dontrun{
query = c("M210903_008_1_1_4704.d",
          "M210903_017_1_1_4713.d",
          "M210903_026_1_1_4722.d")
rawpathfinder(query)
}
}
